Refactor platform specific vdev path stripping
On FreeBSD we only remove "/dev/" while on Linux everything up to and
including the rightmost '/' is removed. The latter would break
expectations about how devices such as /dev/gpt/mygptlabel are shown
on FreeBSD.

Make zfs_strip_partition_path() private to Linux while in the area.
It's only used in the Linux specific file where it's defined.